+++ desc/objectives/_neoclassical.py
@@ -1,15 +1,11 @@
 """Objectives for neoclassical transport."""
 
-import warnings
-
-import numpy as np
-from orthax.legendre import leggauss
-
 from desc.backend import jnp
 from desc.compute import get_profiles, get_transforms
 from desc.compute.utils import _compute as compute_fun
 from desc.grid import LinearGrid
-from desc.integrals._interp_utils import bijection_from_disc, cheb_pts, fourier_pts
+from desc.integrals._bounce_utils import Y_B_rule, num_well_rule
+from desc.integrals.bounce_integral import Bounce2D
 from desc.utils import parse_argname_change, setdefault, warnif
 
 from ..integrals.quad_utils import chebgauss2
@@ -24,8 +20,10 @@ _bounce_overwrite = {
         output of the objective. Has no effect on ``self.grad`` or ``self.hess`` which
         always use reverse mode and forward over reverse mode respectively.
 
-        In ``rev`` mode, reducing the parameter ``pitch_batch_size`` does not
-        reduce memory consumption, so it is recommended to retain the default for that.
+        The default mode of ``auto`` will likely choose ``rev`` for this objective.
+        In ``rev`` mode, reducing the value of the parameter ``pitch_batch_size`` does
+        not reduce memory consumption, so it is recommended to retain the default unless
+        you have explicitly requested to use ``fwd`` mode.
         """
 }
 
@@ -42,15 +40,19 @@ class EffectiveRipple(_Objective):
 
     References
     ----------
-    https://doi.org/10.1063/1.873749.
-    Evaluation of 1/ν neoclassical transport in stellarators.
-    V. V. Nemov, S. V. Kasilov, W. Kernbichler, M. F. Heyn.
-    Phys. Plasmas 1 December 1999; 6 (12): 4622–4632.
+    [1] Evaluation of 1/ν neoclassical transport in stellarators.
+        V. V. Nemov, S. V. Kasilov, W. Kernbichler, M. F. Heyn.
+        Phys. Plasmas 1 December 1999; 6 (12): 4622–4632.
+        https://doi.org/10.1063/1.873749.
+
+    [2] Spectrally accurate, reverse-mode differentiable bounce-averaging
+        algorithm and its applications.
+        Kaya E. Unalmis, Rahul Gaur, Rory Conlin, Dario Panici, Egemen Kolemen.
+        https://arxiv.org/abs/2412.01724.
 
     Notes
     -----
     Performance will improve significantly by resolving these GitHub issues.
-      * https://github.com/jax-ml/jax/issues/30627
       * ``1303`` Patch for differentiable code with dynamic shapes
       * ``1206`` Upsample data above midplane to full grid assuming stellarator symmetry
       * ``1034`` Optimizers/objectives with auxiliary output
@@ -67,17 +69,18 @@ class EffectiveRipple(_Objective):
         Determines the flux surfaces to compute on and resolution of FFTs.
         Default grid samples the boundary surface at ρ=1.
     X : int
-        Poloidal Fourier grid resolution to interpolate the poloidal coordinate.
+        Poloidal Fourier grid resolution to interpolate the angle.
         Preferably rounded down to power of 2.
     Y : int
-        Toroidal Chebyshev grid resolution to interpolate the poloidal coordinate.
+        Toroidal Chebyshev grid resolution over a single field period
+        to interpolate the angle.
         Preferably rounded down to power of 2.
     Y_B : int
         Desired resolution for algorithm to compute bounce points.
-        Default is double ``Y``. Something like 100 is usually sufficient.
-        Currently, this is the number of knots per toroidal transit over
+        Something like 100 is usually sufficient.
+        This is the number of knots per toroidal transit over
         to approximate B with cubic splines.
-    alpha : np.ndarray
+    alpha : jnp.ndarray
         Shape (num alpha, ).
         Starting field line poloidal labels.
         Default is single field line. To compute a surface average
@@ -91,14 +94,14 @@ class EffectiveRipple(_Objective):
         irrational magnetic surface better, with diminishing returns.
     num_well : int
         Maximum number of wells to detect for each pitch and field line.
-        Giving ``None`` will detect all wells but due to current limitations in
+        Giving ``-1`` will detect all wells but due to current limitations in
         JAX this will have worse performance.
         Specifying a number that tightly upper bounds the number of wells will
         increase performance. In general, an upper bound on the number of wells
-        per toroidal transit is ``Aι+B`` where ``A``, ``B`` are the poloidal and
+        per toroidal transit is ``Aι+C`` where ``A``, ``C`` are the poloidal and
         toroidal Fourier resolution of B, respectively, in straight-field line
         PEST coordinates, and ι is the rotational transform normalized by 2π.
-        A tighter upper bound than ``num_well=(Aι+B)*num_transit`` is preferable.
+        A tighter upper bound than ``num_well=(Aι+C)*num_transit`` is preferable.
         The ``check_points`` or ``plot`` methods in ``desc.integrals.Bounce2D``
         are useful to select a reasonable value.
 
@@ -162,7 +165,7 @@ class EffectiveRipple(_Objective):
         X=16,
         Y=32,
         Y_B=None,
-        alpha=np.array([0.0]),
+        alpha=jnp.array([0.0]),
         num_transit=20,
         num_well=None,
         num_quad=32,
@@ -191,23 +194,24 @@ class EffectiveRipple(_Objective):
             use_bounce1d, kwargs, "spline", "use_bounce1d"
         )
         self._grid = grid
-        self._constants = {
-            "quad_weights": 1.0,
-            "alpha": alpha,
-            "X": fourier_pts(X),
-            "Y": cheb_pts(Y, (0, 2 * np.pi))[::-1],
-        }
-        Y_B = setdefault(Y_B, 2 * Y)
+        self._constants = {"quad_weights": 1.0, "alpha": alpha}
         self._hyperparam = {
+            "X": X,
+            "Y": Y,
             "Y_B": Y_B,
             "num_transit": num_transit,
-            "num_well": setdefault(num_well, Y_B * num_transit),
+            "num_well": num_well,
             "num_quad": num_quad,
             "num_pitch": num_pitch,
             "pitch_batch_size": pitch_batch_size,
             "surf_batch_size": surf_batch_size,
             "nufft_eps": nufft_eps,
         }
+        if use_bounce1d:
+            self._hyperparam.pop("X")
+            self._hyperparam.pop("Y")
+            self._hyperparam.pop("pitch_batch_size")
+            self._hyperparam.pop("nufft_eps")
 
         super().__init__(
             things=eq,
@@ -236,36 +240,7 @@ class EffectiveRipple(_Objective):
         if self._use_bounce1d:
             return self._build_bounce1d(use_jit, verbose)
 
-        eq = self.things[0]
-        if self._grid is None:
-            self._grid = LinearGrid(M=eq.M_grid, N=eq.N_grid, NFP=eq.NFP, sym=False)
-        assert self._grid.can_fft2
-
-        rho = self._grid.compress(self._grid.nodes[:, 0])
-        x, w = leggauss(self._hyperparam["Y_B"] // 2)
-        self._constants["_vander"] = _get_vander(self, x)
-        self._constants["fieldline quad"] = (x, w)
-        self._constants["quad"] = chebgauss2(self._hyperparam.pop("num_quad"))
-        self._constants["profiles"] = get_profiles(
-            "effective ripple", eq, grid=self._grid
-        )
-        self._constants["transforms"] = get_transforms(
-            "effective ripple", eq, grid=self._grid
-        )
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", "Unequal number of field periods")
-            # TODO(#1243): Set grid.sym=eq.sym once basis is padded for partial sum
-            self._constants["lambda"] = get_transforms(
-                "lambda",
-                eq,
-                grid=LinearGrid(rho=rho, M=eq.L_basis.M, zeta=self._constants["Y"]),
-            )["L"]
-        assert self._constants["lambda"].basis.NFP == eq.NFP
-
-        self._dim_f = self._grid.num_rho
-        self._target, self._bounds = _parse_callable_target_bounds(
-            self._target, self._bounds, rho
-        )
+        Bounce2D._objective_build(self, names="effective ripple", singular="deriv")
         super().build(use_jit=use_jit, verbose=verbose)
 
     def compute(self, params, constants=None):
@@ -296,12 +271,13 @@ class EffectiveRipple(_Objective):
         data = compute_fun(
             eq, "iota", params, constants["transforms"], constants["profiles"]
         )
-        theta = eq._map_clebsch_coordinates(
-            iota=constants["transforms"]["grid"].compress(data["iota"]),
-            alpha=constants["X"],
-            zeta=constants["Y"],
-            L_lmn=params["L_lmn"],
-            lmbda=constants["lambda"],
+        delta = eq._map_poloidal_coordinates(
+            constants["transforms"]["grid"].compress(data["iota"]),
+            constants["x"],
+            constants["y"],
+            params["L_lmn"],
+            constants["lambda"],
+            outbasis="delta",
             # TODO (#1034): Use old theta values as initial guess.
             tol=1e-7,
         )[..., ::-1]
@@ -313,9 +289,8 @@ class EffectiveRipple(_Objective):
             constants["transforms"],
             constants["profiles"],
             data,
-            theta=theta,
+            angle=delta,
             alpha=constants["alpha"],
-            fieldline_quad=constants["fieldline quad"],
             quad=constants["quad"],
             _vander=constants["_vander"],
             **self._hyperparam,
@@ -323,14 +298,25 @@ class EffectiveRipple(_Objective):
         return constants["transforms"]["grid"].compress(data["effective ripple"])
 
     def _build_bounce1d(self, use_jit=True, verbose=1):
+        eq = self.things[0]
+        if self._grid is None:
+            self._grid = LinearGrid(M=eq.M_grid, N=eq.N_grid, NFP=eq.NFP, sym=eq.sym)
+        assert self._grid.is_meshgrid and eq.sym == self._grid.sym
+
         Y_B = self._hyperparam.pop("Y_B")
+        Y_B = setdefault(Y_B, Y_B_rule(32, eq.NFP, spline=True))
+
         num_transit = self._hyperparam.pop("num_transit")
+
+        if self._hyperparam["num_well"] is None:
+            self._hyperparam["num_well"] = num_well_rule(num_transit, eq.NFP, Y_B)
+
         num_quad = self._hyperparam.pop("num_quad")
-        self._hyperparam.pop("nufft_eps")
-        del self._constants["X"]
-        self._constants["Y"] = np.linspace(
-            0, 2 * np.pi * num_transit, Y_B * num_transit
+
+        self._constants["zeta"] = jnp.linspace(
+            0, 2 * jnp.pi * num_transit, Y_B * num_transit
         )
+
         self._keys_1dr = [
             "iota",
             "iota_r",
@@ -339,11 +325,6 @@ class EffectiveRipple(_Objective):
             "max_tz |B|",
             "R0",
         ]
-
-        eq = self.things[0]
-        if self._grid is None:
-            self._grid = LinearGrid(M=eq.M_grid, N=eq.N_grid, NFP=eq.NFP, sym=eq.sym)
-        assert self._grid.is_meshgrid and eq.sym == self._grid.sym
 
         rho = self._grid.compress(self._grid.nodes[:, 0])
         self._constants["rho"] = rho
@@ -373,12 +354,10 @@ class EffectiveRipple(_Objective):
             constants["transforms_1dr"],
             constants["profiles"],
         )
-        # TODO(#1243): Upgrade this to use _map_clebsch_coordinates once
-        #  the note in _L_partial_sum method is resolved.
         grid = eq._get_rtz_grid(
             constants["rho"],
             constants["alpha"],
-            constants["Y"],
+            constants["zeta"],
             coordinates="raz",
             iota=self._grid.compress(data["iota"]),
             params=params,
@@ -402,25 +381,3 @@ class EffectiveRipple(_Objective):
             **self._hyperparam,
         )
         return grid.compress(data["old effective ripple"])
-
-
-def _get_vander(obj, x):
-    eq = obj.things[0]
-    Y_B = ((obj._hyperparam["Y_B"] + eq.NFP - 1) // eq.NFP) * eq.NFP
-    return {
-        "dct cfl": _vander_dct_cfl(x, obj._constants["Y"].size),
-        "dft cfl": _vander_dft_cfl(x, obj._grid),
-        "dct spline": _vander_dct_cfl(
-            jnp.linspace(-1, 1, Y_B, endpoint=False), obj._constants["Y"].size
-        ),
-    }
-
-
-def _vander_dft_cfl(x, grid):
-    modes = jnp.fft.fftfreq(grid.num_zeta, 1 / (grid.NFP * grid.num_zeta))
-    zeta = bijection_from_disc(x, 0, 2 * jnp.pi)
-    return jnp.exp(1j * modes * zeta[:, jnp.newaxis])[..., jnp.newaxis]
-
-
-def _vander_dct_cfl(x, Y):
-    return jnp.cos(jnp.arange(Y) * jnp.arccos(x)[:, jnp.newaxis])